Partition vector algorithms test: move out lex compare family
