+++ tests/std/test.lst
@@ -724,6 +724,7 @@ tests\VSO_0000000_string_view_idl
 tests\VSO_0000000_type_traits
 tests\VSO_0000000_vector_algorithms
 tests\VSO_0000000_vector_algorithms_floats
+tests\VSO_0000000_vector_algorithms_mismatch_and_lex_compare
 tests\VSO_0000000_wcfb01_idempotent_container_destructors
 tests\VSO_0000000_wchar_t_filebuf_xsmeown
 tests\VSO_0095468_clr_exception_ptr_bad_alloc
