+++ tests/std/tests/VSO_0000000_vector_algorithms_mismatch_and_lex_compare/env.lst
@@ -1,0 +1,7 @@
+# Copyright (c) Microsoft Corporation.
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+RUNALL_INCLUDE ..\usual_matrix.lst
+RUNALL_CROSSLIST
+*	PM_CL=""
+*	PM_CL="/D_USE_STD_VECTOR_ALGORITHMS=0"
